+++ packages/components/dummy/src/Dummy.stories.tsx
@@ -1,9 +1,0 @@
-import { ReactLiveBlock } from '@docs/helpers/ReactLiveBlock'
-
-import { Dummy } from '.'
-
-export const Default = () => (
-  <ReactLiveBlock scope={{ Dummy }}>
-    <Dummy type="foo" />
-  </ReactLiveBlock>
-)