+++ packages/components/radio/src/RadioGroupContext.tsx
@@ -1,0 +1,5 @@
+import { createContext } from 'react'
+
+import { RadioVariantsProps } from './Radio.variants'
+
+export const RadioGroupContext = createContext<Pick<RadioVariantsProps, 'intent' | 'size'>>({})