+++ packages/utils/cli/bin/spark.mjs
@@ -8,7 +8,7 @@ const { version } = require('../package.json')
 
 program.version(version, '--version')
 
-program.command('generate <component>', 'Generate a component scaffolding').alias('g')
+program.command('generate', 'Generate a component scaffolding').alias('g')
 program.command('setup-themes', 'Set up Spark theming configuration')
 
 program.parse(process.argv)