+++ packages/utils/tailwind-plugins/index.js
@@ -1,8 +1,10 @@
 /* eslint-disable @typescript-eslint/no-var-requires */
 const animations = require('./animations')
 const sizings = require('./sizings')
+const sparkTheme = require('./spark-theme')
 
 module.exports = {
   animations,
   sizings,
+  sparkTheme,
 }