+++ packages/utils/tailwind-plugins/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.2](https://github.com/adevinta/spark/compare/@spark-ui/tailwind-plugins@2.1.1...@spark-ui/tailwind-plugins@2.1.2) (2023-03-20)
+
+### Bug Fixes
+
+- **tailwind-plugins:** fix invalid tailwind package version ([b79c5a6](https://github.com/adevinta/spark/commit/b79c5a6e92c955246a57728dfb466d0d23d82e3d))
+
+## [2.1.1](https://github.com/adevinta/spark/compare/@spark-ui/tailwind-plugins@2.1.0...@spark-ui/tailwind-plugins@2.1.1) (2023-03-19)
+
+**Note:** Version bump only for package @spark-ui/tailwind-plugins
+
+# [2.1.0](https://github.com/adevinta/spark/compare/@spark-ui/tailwind-plugins@2.0.0...@spark-ui/tailwind-plugins@2.1.0) (2023-03-17)
+
+### Bug Fixes
+
+- add early return ([3cc0916](https://github.com/adevinta/spark/commit/3cc0916d4fe1aafd76814646b804be45b8870f34)), closes [#411](https://github.com/adevinta/spark/issues/411)
+
+### Features
+
+- **tailwind-plugins:** add spark theme plugin ([7009c51](https://github.com/adevinta/spark/commit/7009c513ab37b1118b89d7bada365156fb85f362)), closes [#411](https://github.com/adevinta/spark/issues/411)
+
 # [2.0.0](https://github.com/adevinta/spark/compare/@spark-ui/tailwind-plugins@1.1.2...@spark-ui/tailwind-plugins@2.0.0) (2023-03-15)
 
 ### Features